Move URLSessionHTTPClient to its own file in production
