+++ MyEssentialFeedTests/Feed API/URLSessionHTTPClientTests.swift
@@ -9,28 +9,7 @@ import XCTest
 import MyEssentialFeed
 
 
-class URLSessionHTTPClient: HTTPClient{
-    private let session: URLSession
-    
-    init(session: URLSession = .shared){
-        self.session = session
-    }
-    struct UnexpectedValuesRepresntation: Error{
-        
-    }
-    func get(from url: URL, compoletion completion: @escaping (HTTPClientResult)->Void){
-        session.dataTask(with: url){data, response, error in
-            if let error = error{
-                completion(.failure(error))
-            }else if let data = data, let response = response as? HTTPURLResponse{
-                completion(.success(data, response))
-            }else{
-                completion(.failure(UnexpectedValuesRepresntation()))
-            }
-        }
-        .resume()
-    }
-}
+
 
 class URLSessionHTTPClientTests: XCTestCase {
